+++ components/cdc/src/delegate.rs
@@ -1,14 +1,12 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::collections::BTreeSet;
+use std::cell::RefCell;
 use std::mem;
-use std::ops::Deref;
+use std::rc::Rc;
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 use std::sync::Arc;
 
 use crossbeam::atomic::AtomicCell;
-use engine_rocks::{RocksEngine, RocksSnapshot};
-use engine_traits::{IterOptions, KvEngine, ReadOptions, CF_DEFAULT, CF_LOCK, CF_WRITE};
 #[cfg(feature = "prost-codec")]
 use kvproto::cdcpb::{
     error::DuplicateRequest as ErrorDuplicateRequest,
@@ -25,23 +23,19 @@ use kvproto::cdcpb::{
 };
 use kvproto::errorpb;
 use kvproto::kvrpcpb::ExtraOp as TxnExtraOp;
-use kvproto::metapb::{Peer, Region, RegionEpoch};
+use kvproto::metapb::{Region, RegionEpoch};
 use kvproto::raft_cmdpb::{AdminCmdType, AdminRequest, AdminResponse, CmdType, Request};
 use raftstore::coprocessor::{Cmd, CmdBatch};
 use raftstore::store::fsm::ObserveID;
 use raftstore::store::util::compare_region_epoch;
-use raftstore::store::RegionSnapshot;
 use raftstore::Error as RaftStoreError;
 use resolved_ts::Resolver;
 use tikv::storage::txn::TxnEntry;
-use tikv::storage::{Cursor, ScanMode, Snapshot as EngineSnapshot, Statistics};
 use tikv_util::collections::HashMap;
 use tikv_util::mpsc::batch::Sender as BatchSender;
-use tikv_util::Either;
-use txn_types::{
-    Key, Lock, LockType, MutationType, TimeStamp, TxnExtra, Value, WriteRef, WriteType,
-};
+use txn_types::{Key, Lock, LockType, TimeStamp, WriteRef, WriteType};
 
+use crate::endpoint::OldValueCallback;
 use crate::metrics::*;
 use crate::service::ConnID;
 use crate::{Error, Result};
@@ -408,8 +402,7 @@ impl Delegate {
     pub fn on_batch(
         &mut self,
         batch: CmdBatch,
-        txn_extra: &mut TxnExtra,
-        kv_engine: &RocksEngine,
+        old_value_cb: Rc<RefCell<OldValueCallback>>,
     ) -> Result<()> {
         // Stale CmdBatch, drop it sliently.
         if batch.observe_id != self.id {
@@ -423,7 +416,7 @@ impl Delegate {
             } = cmd;
             if !response.get_header().has_error() {
                 if !request.has_admin_request() {
-                    self.sink_data(index, request.requests.into(), txn_extra, kv_engine)?;
+                    self.sink_data(index, request.requests.into(), old_value_cb.clone())?;
                 } else {
                     self.sink_admin(request.take_admin_request(), response.take_admin_response())?;
                 }
@@ -535,13 +528,10 @@ impl Delegate {
         &mut self,
         index: u64,
         requests: Vec<Request>,
-        txn_extra: &mut TxnExtra,
-        kv_engine: &RocksEngine,
+        old_value_cb: Rc<RefCell<OldValueCallback>>,
     ) -> Result<()> {
         let mut rows = HashMap::default();
         let mut total_size = 0;
-        let mut to_seek_old = BTreeSet::new();
-        let mut old_value_reader = None;
         for mut req in requests {
             // CDC cares about put requests only.
             if req.get_cmd_type() != CmdType::Put {
@@ -602,32 +592,8 @@ impl Delegate {
                     }
 
                     if self.txn_extra_op == TxnExtraOp::ReadOldValue {
-                        old_value_reader = old_value_reader.or_else(|| {
-                            // Hack: init a empty region to pass the check
-                            let mut region = Region::default();
-                            region.mut_peers().push(Peer::default());
-                            let snapshot = RegionSnapshot::from_snapshot(
-                                Arc::new(kv_engine.snapshot()),
-                                region,
-                            );
-                            Some(OldValueReader::new(snapshot))
-                        });
-                        match self.old_value_from_extra(
-                            &row,
-                            txn_extra,
-                            old_value_reader.as_mut().unwrap(),
-                        ) {
-                            Either::Left(key) => {
-                                to_seek_old.insert(key);
-                            }
-                            Either::Right(Some(value)) => row.old_value = value,
-                            Either::Right(None) => {
-                                // TODO(5kbpers): may add a log here.
-                                self.mark_failed();
-                                let store_err = RaftStoreError::RegionNotFound(self.region_id);
-                                return Err(Error::Request(store_err.into()));
-                            }
-                        }
+                        let key = Key::from_raw(&row.key).append_ts(row.start_ts.into());
+                        row.old_value = old_value_cb.borrow_mut()(key).unwrap_or_default();
                     }
 
                     let occupied = rows.entry(row.key.clone()).or_default();
@@ -669,9 +635,6 @@ impl Delegate {
                 }
             }
         }
-        if !to_seek_old.is_empty() {
-            self.seek_for_old_value(to_seek_old, &mut rows, old_value_reader.as_mut().unwrap())?;
-        }
         let mut entries = Vec::with_capacity(rows.len());
         for (_, v) in rows {
             entries.push(v);
@@ -708,163 +671,6 @@ impl Delegate {
         };
         self.mark_failed();
         Err(Error::Request(store_err.into()))
-    }
-
-    // Return Either::Right(Some(...)) if key was contained in the extra and can get it from short_value or default cf.
-    // Return Either::Right(None) if key was contained in the extra and cannot get it from short_value and default cf.
-    // Return Either::Left(key) if key wasn't contained in the extra, needs to be seeked from the engine.
-    fn old_value_from_extra(
-        &mut self,
-        row: &EventRow,
-        txn_extra: &mut TxnExtra,
-        reader: &mut OldValueReader<RegionSnapshot<RocksSnapshot>>,
-    ) -> Either<Key, Option<Value>> {
-        let old_values = txn_extra.mut_old_values();
-        let key = Key::from_raw(row.key.as_slice()).append_ts(row.start_ts.into());
-        if let Some((old_value, mutation_type)) = old_values.remove(&key) {
-            match mutation_type {
-                MutationType::Insert => assert!(old_value.is_none()),
-                MutationType::Put | MutationType::Delete => {
-                    if let Some(old_value) = old_value {
-                        let start_ts = old_value.start_ts;
-                        return Either::Right(old_value.short_value.or_else(|| {
-                            let prev_key = key.truncate_ts().unwrap().append_ts(start_ts);
-                            let mut opts = ReadOptions::new();
-                            opts.set_fill_cache(false);
-                            reader.get_value_default(&prev_key)
-                        }));
-                    }
-                }
-                _ => unreachable!(),
-            }
-        }
-        Either::Left(key)
-    }
-
-    // to_seek_old contains encoded keys.
-    fn seek_for_old_value(
-        &mut self,
-        to_seek_old: BTreeSet<Key>,
-        rows: &mut HashMap<Vec<u8>, EventRow>,
-        reader: &mut OldValueReader<RegionSnapshot<RocksSnapshot>>,
-    ) -> Result<()> {
-        for key in to_seek_old {
-            match reader.near_seek_old_value(&key)? {
-                Some(value) => {
-                    let raw_key = key.truncate_ts().unwrap().into_raw().unwrap();
-                    rows.get_mut(&raw_key).unwrap().old_value = value;
-                }
-                None => {
-                    self.mark_failed();
-                    let store_err = RaftStoreError::RegionNotFound(self.region_id);
-                    return Err(Error::Request(store_err.into()));
-                }
-            }
-        }
-        Ok(())
-    }
-}
-
-struct OldValueReader<S: EngineSnapshot> {
-    snapshot: S,
-    write_cursor: Cursor<S::Iter>,
-    // TODO(5kbpers): add a metric here.
-    statistics: Statistics,
-}
-
-impl<S: EngineSnapshot> OldValueReader<S> {
-    fn new(snapshot: S) -> Self {
-        let mut iter_opts = IterOptions::default()
-            .use_prefix_seek()
-            .set_prefix_same_as_start(true);
-        iter_opts.set_fill_cache(false);
-        let write_cursor = snapshot
-            .iter_cf(CF_WRITE, iter_opts, ScanMode::Mixed)
-            .unwrap();
-        Self {
-            snapshot,
-            write_cursor,
-            statistics: Statistics::default(),
-        }
-    }
-
-    // return Some(vec![]) if value is empty.
-    // return None if key not exist.
-    fn get_value_default(&mut self, key: &Key) -> Option<Value> {
-        self.statistics.data.get += 1;
-        let mut opts = ReadOptions::new();
-        opts.set_fill_cache(false);
-        self.snapshot
-            .get_cf_opt(opts, CF_DEFAULT, &key)
-            .unwrap()
-            .map(|v| v.deref().to_vec())
-    }
-
-    fn check_lock(&mut self, key: &Key) -> bool {
-        self.statistics.lock.get += 1;
-        let mut opts = ReadOptions::new();
-        opts.set_fill_cache(false);
-        let key_slice = key.as_encoded();
-        let user_key = Key::from_encoded_slice(Key::truncate_ts_for(key_slice).unwrap());
-
-        match self.snapshot.get_cf_opt(opts, CF_LOCK, &user_key).unwrap() {
-            Some(v) => {
-                let lock = Lock::parse(v.deref()).unwrap();
-                lock.ts == Key::decode_ts_from(key_slice).unwrap()
-            }
-            None => false,
-        }
-    }
-
-    // return Some(vec![]) if value is empty.
-    // return None if key not exist.
-    fn near_seek_old_value(&mut self, key: &Key) -> Result<Option<Value>> {
-        let user_key = Key::truncate_ts_for(key.as_encoded()).unwrap();
-        if self
-            .write_cursor
-            .near_seek(key, &mut self.statistics.write)?
-            && Key::is_user_key_eq(self.write_cursor.key(&mut self.statistics.write), user_key)
-        {
-            if self.write_cursor.key(&mut self.statistics.write) == key.as_encoded().as_slice() {
-                // Key was committed, move cursor to the next key to seek for old value.
-                if !self.write_cursor.next(&mut self.statistics.write) {
-                    // Do not has any next key, return empty value.
-                    return Ok(Some(Vec::default()));
-                }
-            } else if !self.check_lock(key) {
-                return Ok(None);
-            }
-
-            // Key was not committed, check if the lock is corresponding to the key.
-            let mut old_value = Some(Vec::default());
-            while Key::is_user_key_eq(self.write_cursor.key(&mut self.statistics.write), user_key) {
-                let write =
-                    WriteRef::parse(self.write_cursor.value(&mut self.statistics.write)).unwrap();
-                old_value = match write.write_type {
-                    WriteType::Put => match write.short_value {
-                        Some(short_value) => Some(short_value.to_vec()),
-                        None => {
-                            let key = key.clone().truncate_ts().unwrap().append_ts(write.start_ts);
-                            self.get_value_default(&key)
-                        }
-                    },
-                    WriteType::Delete => Some(Vec::default()),
-                    WriteType::Rollback | WriteType::Lock => {
-                        if !self.write_cursor.next(&mut self.statistics.write) {
-                            Some(Vec::default())
-                        } else {
-                            continue;
-                        }
-                    }
-                };
-                break;
-            }
-            Ok(old_value)
-        } else if self.check_lock(key) {
-            Ok(Some(Vec::default()))
-        } else {
-            Ok(None)
-        }
     }
 }
 
@@ -947,9 +753,7 @@ mod tests {
     use kvproto::errorpb::Error as ErrorHeader;
     use kvproto::metapb::Region;
     use std::cell::Cell;
-    use tikv::storage::kv::TestEngineBuilder;
     use tikv::storage::mvcc::test_util::*;
-    use tikv::storage::mvcc::tests::*;
     use tikv_util::mpsc::batch::{self, BatchReceiver, VecCollector};
 
     #[test]
@@ -1181,55 +985,5 @@ mod tests {
         let mut resolver = Resolver::new(region_id);
         resolver.init();
         delegate.on_region_ready(resolver, region);
-    }
-
-    #[test]
-    fn test_old_value_reader() {
-        let engine = TestEngineBuilder::new().build().unwrap();
-        let kv_engine = engine.get_rocksdb();
-        let k = b"k";
-        let key = Key::from_raw(k);
-
-        let must_get_eq = |ts: u64, value| {
-            let mut old_value_reader = OldValueReader::new(Arc::new(kv_engine.snapshot()));
-            assert_eq!(
-                old_value_reader
-                    .near_seek_old_value(&key.clone().append_ts(ts.into()))
-                    .unwrap(),
-                value
-            );
-            let mut opts = ReadOptions::new();
-            opts.set_fill_cache(false);
-            // engine
-            //     .get_rocksdb()
-            //     .snapshot()
-            //     .get_value_cf_opt(&opts, CF_LOCK, key.as_encoded())
-            //     .unwrap()
-        };
-
-        must_prewrite_put(&engine, k, b"v1", k, 1);
-        must_get_eq(2, None);
-        must_get_eq(1, Some(vec![]));
-        must_commit(&engine, k, 1, 1);
-        must_get_eq(1, Some(vec![]));
-
-        must_prewrite_put(&engine, k, b"v2", k, 2);
-        must_get_eq(2, Some(b"v1".to_vec()));
-        must_rollback(&engine, k, 2);
-
-        must_prewrite_put(&engine, k, b"v3", k, 3);
-        must_get_eq(3, Some(b"v1".to_vec()));
-        must_commit(&engine, k, 3, 3);
-
-        must_prewrite_delete(&engine, k, k, 4);
-        must_get_eq(4, Some(b"v3".to_vec()));
-        must_commit(&engine, k, 4, 4);
-
-        must_prewrite_put(&engine, k, vec![b'v'; 5120].as_slice(), k, 5);
-        must_get_eq(5, Some(vec![]));
-        must_commit(&engine, k, 5, 5);
-
-        must_prewrite_delete(&engine, k, k, 6);
-        must_get_eq(6, Some(vec![b'v'; 5120]));
     }
 }
