+++ components/cdc/src/observer.rs
@@ -1,19 +1,24 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
 use std::cell::RefCell;
+use std::ops::Deref;
 use std::sync::{Arc, RwLock};
 
 use engine_rocks::RocksEngine;
+use engine_traits::{IterOptions, KvEngine, ReadOptions, CF_DEFAULT, CF_LOCK, CF_WRITE};
+use kvproto::metapb::{Peer, Region};
 use raft::StateRole;
 use raftstore::coprocessor::*;
 use raftstore::store::fsm::ObserveID;
+use raftstore::store::RegionSnapshot;
 use raftstore::Error as RaftStoreError;
+use tikv::storage::{Cursor, ScanMode, Snapshot as EngineSnapshot, Statistics};
 use tikv_util::collections::HashMap;
 use tikv_util::worker::Scheduler;
-use txn_types::TxnExtra;
+use txn_types::{Key, Lock, MutationType, TxnExtra, Value, WriteRef, WriteType};
 
 use crate::endpoint::{Deregister, Task};
-use crate::Error as CdcError;
+use crate::{Error as CdcError, Result};
 
 /// An Observer for CDC.
 ///
@@ -92,7 +97,7 @@ impl CdcObserver {
 
 impl Coprocessor for CdcObserver {}
 
-impl CmdObserver for CdcObserver {
+impl<E: KvEngine> CmdObserver<E> for CdcObserver {
     fn on_prepare_for_apply(&self, observe_id: ObserveID, region_id: u64) {
         self.cmd_batches
             .borrow_mut()
@@ -107,13 +112,44 @@ impl CmdObserver for CdcObserver {
             .push(observe_id, region_id, cmd);
     }
 
-    fn on_flush_apply(&self, txn_extras: Vec<TxnExtra>) {
+    fn on_flush_apply(&self, txn_extras: Vec<TxnExtra>, engine: E) {
         fail_point!("before_cdc_flush_apply");
+        let mut txn_extra = TxnExtra::default();
+        txn_extras
+            .into_iter()
+            .for_each(|mut e| txn_extra.extend(&mut e));
         if !self.cmd_batches.borrow().is_empty() {
             let batches = self.cmd_batches.replace(Vec::default());
+            let mut region = Region::default();
+            region.mut_peers().push(Peer::default());
+            let snapshot = RegionSnapshot::from_snapshot(Arc::new(engine.snapshot()), region);
+            let mut reader = OldValueReader::new(snapshot);
+            let get_old_value = move |key| {
+                if let Some((old_value, mutation_type)) = txn_extra.mut_old_values().remove(&key) {
+                    match mutation_type {
+                        MutationType::Insert => {
+                            assert!(old_value.is_none());
+                            return None;
+                        }
+                        MutationType::Put | MutationType::Delete => {
+                            if let Some(old_value) = old_value {
+                                let start_ts = old_value.start_ts;
+                                return old_value.short_value.or_else(|| {
+                                    let prev_key = key.truncate_ts().unwrap().append_ts(start_ts);
+                                    let mut opts = ReadOptions::new();
+                                    opts.set_fill_cache(false);
+                                    reader.get_value_default(&prev_key)
+                                });
+                            }
+                        }
+                        _ => unreachable!(),
+                    }
+                }
+                reader.near_seek_old_value(&key).unwrap_or_default()
+            };
             if let Err(e) = self.sched.schedule(Task::MultiBatch {
                 multi: batches,
-                txn_extras,
+                old_value_cb: Box::new(get_old_value),
             }) {
                 warn!("schedule cdc task failed"; "error" => ?e);
             }
@@ -166,26 +202,134 @@ impl RegionChangeObserver for CdcObserver {
     }
 }
 
+struct OldValueReader<S: EngineSnapshot> {
+    snapshot: S,
+    write_cursor: Cursor<S::Iter>,
+    // TODO(5kbpers): add a metric here.
+    statistics: Statistics,
+}
+
+impl<S: EngineSnapshot> OldValueReader<S> {
+    fn new(snapshot: S) -> Self {
+        let mut iter_opts = IterOptions::default()
+            .use_prefix_seek()
+            .set_prefix_same_as_start(true);
+        iter_opts.set_fill_cache(false);
+        let write_cursor = snapshot
+            .iter_cf(CF_WRITE, iter_opts, ScanMode::Mixed)
+            .unwrap();
+        Self {
+            snapshot,
+            write_cursor,
+            statistics: Statistics::default(),
+        }
+    }
+
+    // return Some(vec![]) if value is empty.
+    // return None if key not exist.
+    fn get_value_default(&mut self, key: &Key) -> Option<Value> {
+        self.statistics.data.get += 1;
+        let mut opts = ReadOptions::new();
+        opts.set_fill_cache(false);
+        self.snapshot
+            .get_cf_opt(opts, CF_DEFAULT, &key)
+            .unwrap()
+            .map(|v| v.deref().to_vec())
+    }
+
+    fn check_lock(&mut self, key: &Key) -> bool {
+        self.statistics.lock.get += 1;
+        let mut opts = ReadOptions::new();
+        opts.set_fill_cache(false);
+        let key_slice = key.as_encoded();
+        let user_key = Key::from_encoded_slice(Key::truncate_ts_for(key_slice).unwrap());
+
+        match self.snapshot.get_cf_opt(opts, CF_LOCK, &user_key).unwrap() {
+            Some(v) => {
+                let lock = Lock::parse(v.deref()).unwrap();
+                lock.ts == Key::decode_ts_from(key_slice).unwrap()
+            }
+            None => false,
+        }
+    }
+
+    // return Some(vec![]) if value is empty.
+    // return None if key not exist.
+    fn near_seek_old_value(&mut self, key: &Key) -> Result<Option<Value>> {
+        let user_key = Key::truncate_ts_for(key.as_encoded()).unwrap();
+        if self
+            .write_cursor
+            .near_seek(key, &mut self.statistics.write)?
+            && Key::is_user_key_eq(self.write_cursor.key(&mut self.statistics.write), user_key)
+        {
+            if self.write_cursor.key(&mut self.statistics.write) == key.as_encoded().as_slice() {
+                // Key was committed, move cursor to the next key to seek for old value.
+                if !self.write_cursor.next(&mut self.statistics.write) {
+                    // Do not has any next key, return empty value.
+                    return Ok(Some(Vec::default()));
+                }
+            } else if !self.check_lock(key) {
+                return Ok(None);
+            }
+
+            // Key was not committed, check if the lock is corresponding to the key.
+            let mut old_value = Some(Vec::default());
+            while Key::is_user_key_eq(self.write_cursor.key(&mut self.statistics.write), user_key) {
+                let write =
+                    WriteRef::parse(self.write_cursor.value(&mut self.statistics.write)).unwrap();
+                old_value = match write.write_type {
+                    WriteType::Put => match write.short_value {
+                        Some(short_value) => Some(short_value.to_vec()),
+                        None => {
+                            let key = key.clone().truncate_ts().unwrap().append_ts(write.start_ts);
+                            self.get_value_default(&key)
+                        }
+                    },
+                    WriteType::Delete => Some(Vec::default()),
+                    WriteType::Rollback | WriteType::Lock => {
+                        if !self.write_cursor.next(&mut self.statistics.write) {
+                            Some(Vec::default())
+                        } else {
+                            continue;
+                        }
+                    }
+                };
+                break;
+            }
+            Ok(old_value)
+        } else if self.check_lock(key) {
+            Ok(Some(Vec::default()))
+        } else {
+            Ok(None)
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use engine_rocks::RocksEngine;
     use kvproto::metapb::Region;
     use kvproto::raft_cmdpb::*;
     use std::time::Duration;
+    use tikv::storage::kv::TestEngineBuilder;
+    use tikv::storage::mvcc::tests::*;
 
     #[test]
     fn test_register_and_deregister() {
         let (scheduler, rx) = tikv_util::worker::dummy_scheduler();
         let observer = CdcObserver::new(scheduler);
         let observe_id = ObserveID::new();
+        let engine = TestEngineBuilder::new().build().unwrap().get_rocksdb();
 
-        observer.on_prepare_for_apply(observe_id, 0);
-        observer.on_apply_cmd(
+        <CdcObserver as CmdObserver<RocksEngine>>::on_prepare_for_apply(&observer, observe_id, 0);
+        <CdcObserver as CmdObserver<RocksEngine>>::on_apply_cmd(
+            &observer,
             observe_id,
             0,
             Cmd::new(0, RaftCmdRequest::default(), RaftCmdResponse::default()),
         );
-        observer.on_flush_apply(Vec::default());
+        observer.on_flush_apply(Vec::default(), engine);
 
         match rx.recv_timeout(Duration::from_millis(10)).unwrap().unwrap() {
             Task::MultiBatch { multi, .. } => {
@@ -236,5 +380,50 @@ mod tests {
         let mut ctx = ObserverContext::new(&region);
         observer.on_role_change(&mut ctx, StateRole::Follower);
         rx.recv_timeout(Duration::from_millis(10)).unwrap_err();
+    }
+
+    #[test]
+    fn test_old_value_reader() {
+        let engine = TestEngineBuilder::new().build().unwrap();
+        let kv_engine = engine.get_rocksdb();
+        let k = b"k";
+        let key = Key::from_raw(k);
+
+        let must_get_eq = |ts: u64, value| {
+            let mut old_value_reader = OldValueReader::new(Arc::new(kv_engine.snapshot()));
+            assert_eq!(
+                old_value_reader
+                    .near_seek_old_value(&key.clone().append_ts(ts.into()))
+                    .unwrap(),
+                value
+            );
+            let mut opts = ReadOptions::new();
+            opts.set_fill_cache(false);
+        };
+
+        must_prewrite_put(&engine, k, b"v1", k, 1);
+        must_get_eq(2, None);
+        must_get_eq(1, Some(vec![]));
+        must_commit(&engine, k, 1, 1);
+        must_get_eq(1, Some(vec![]));
+
+        must_prewrite_put(&engine, k, b"v2", k, 2);
+        must_get_eq(2, Some(b"v1".to_vec()));
+        must_rollback(&engine, k, 2);
+
+        must_prewrite_put(&engine, k, b"v3", k, 3);
+        must_get_eq(3, Some(b"v1".to_vec()));
+        must_commit(&engine, k, 3, 3);
+
+        must_prewrite_delete(&engine, k, k, 4);
+        must_get_eq(4, Some(b"v3".to_vec()));
+        must_commit(&engine, k, 4, 4);
+
+        must_prewrite_put(&engine, k, vec![b'v'; 5120].as_slice(), k, 5);
+        must_get_eq(5, Some(vec![]));
+        must_commit(&engine, k, 5, 5);
+
+        must_prewrite_delete(&engine, k, k, 6);
+        must_get_eq(6, Some(vec![b'v'; 5120]));
     }
 }
