+++ components/raftstore/Cargo.toml
@@ -91,3 +91,4 @@ git = "https://github.com/tikv/yatp.git"
 [dev-dependencies]
 panic_hook = { path = "../panic_hook" }
 test_sst_importer = { path = "../test_sst_importer" }
+engine_panic = { path = "../engine_panic" }
