+++ components/raftstore/src/coprocessor/dispatcher.rs
@@ -145,7 +145,7 @@ impl_box_observer!(
     RegionChangeObserver,
     WrappedRegionChangeObserver
 );
-impl_box_observer!(BoxCmdObserver, CmdObserver, WrappedCmdObserver);
+impl_box_observer_g!(BoxCmdObserver, CmdObserver, WrappedCmdObserver);
 
 /// Registry contains all registered coprocessors.
 #[derive(Clone)]
@@ -159,7 +159,7 @@ where
     split_check_observers: Vec<Entry<BoxSplitCheckObserver<E>>>,
     role_observers: Vec<Entry<BoxRoleObserver>>,
     region_change_observers: Vec<Entry<BoxRegionChangeObserver>>,
-    cmd_observers: Vec<Entry<BoxCmdObserver>>,
+    cmd_observers: Vec<Entry<BoxCmdObserver<E>>>,
     // TODO: add endpoint
 }
 
@@ -219,7 +219,7 @@ impl<E> Registry<E> {
         push!(priority, rlo, self.region_change_observers);
     }
 
-    pub fn register_cmd_observer(&mut self, priority: u32, rlo: BoxCmdObserver) {
+    pub fn register_cmd_observer(&mut self, priority: u32, rlo: BoxCmdObserver<E>) {
         push!(priority, rlo, self.cmd_observers);
     }
 }
@@ -469,7 +469,7 @@ where
             .on_apply_cmd(observe_id, region_id, cmd)
     }
 
-    pub fn on_flush_apply(&self, txn_extras: Vec<TxnExtra>) {
+    pub fn on_flush_apply(&self, txn_extras: Vec<TxnExtra>, engine: E) {
         if self.registry.cmd_observers.is_empty() {
             return;
         }
@@ -481,7 +481,7 @@ where
                 .unwrap()
                 .observer
                 .inner()
-                .on_flush_apply(txn_extras.clone())
+                .on_flush_apply(txn_extras.clone(), engine.clone())
         }
         self.registry
             .cmd_observers
@@ -489,7 +489,7 @@ where
             .unwrap()
             .observer
             .inner()
-            .on_flush_apply(txn_extras)
+            .on_flush_apply(txn_extras, engine)
     }
 
     pub fn shutdown(&self) {
@@ -514,7 +514,7 @@ mod tests {
     use std::sync::atomic::*;
     use std::sync::Arc;
 
-    use engine_rocks::RocksEngine;
+    use engine_panic::PanicEngine;
     use kvproto::metapb::Region;
     use kvproto::raft_cmdpb::{
         AdminRequest, AdminResponse, RaftCmdRequest, RaftCmdResponse, Request, Response,
@@ -615,14 +615,14 @@ mod tests {
         }
     }
 
-    impl CmdObserver for TestCoprocessor {
+    impl CmdObserver<PanicEngine> for TestCoprocessor {
         fn on_prepare_for_apply(&self, _: ObserveID, _: u64) {
             self.called.fetch_add(11, Ordering::SeqCst);
         }
         fn on_apply_cmd(&self, _: ObserveID, _: u64, _: Cmd) {
             self.called.fetch_add(12, Ordering::SeqCst);
         }
-        fn on_flush_apply(&self, _: Vec<TxnExtra>) {
+        fn on_flush_apply(&self, _: Vec<TxnExtra>, _: PanicEngine) {
             self.called.fetch_add(13, Ordering::SeqCst);
         }
     }
@@ -645,7 +645,7 @@ mod tests {
 
     #[test]
     fn test_trigger_right_hook() {
-        let mut host = CoprocessorHost::<RocksEngine>::default();
+        let mut host = CoprocessorHost::<PanicEngine>::default();
         let ob = TestCoprocessor::default();
         host.registry
             .register_admin_observer(1, BoxAdminObserver::new(ob.clone()));
@@ -701,13 +701,13 @@ mod tests {
             Cmd::new(0, RaftCmdRequest::default(), query_resp),
         );
         assert_all!(&[&ob.called], &[78]);
-        host.on_flush_apply(Vec::default());
+        host.on_flush_apply(Vec::default(), PanicEngine);
         assert_all!(&[&ob.called], &[91]);
     }
 
     #[test]
     fn test_order() {
-        let mut host = CoprocessorHost::<RocksEngine>::default();
+        let mut host = CoprocessorHost::<PanicEngine>::default();
 
         let ob1 = TestCoprocessor::default();
         host.registry
