+++ components/raftstore/src/coprocessor/mod.rs
@@ -238,11 +238,11 @@ impl CmdBatch {
     }
 }
 
-pub trait CmdObserver: Coprocessor {
+pub trait CmdObserver<E>: Coprocessor {
     /// Hook to call after preparing for applying write requests.
     fn on_prepare_for_apply(&self, observe_id: ObserveID, region_id: u64);
     /// Hook to call after applying a write request.
     fn on_apply_cmd(&self, observe_id: ObserveID, region_id: u64, cmd: Cmd);
     /// Hook to call after flushing writes to db.
-    fn on_flush_apply(&self, txn_extras: Vec<TxnExtra>);
+    fn on_flush_apply(&self, txn_extras: Vec<TxnExtra>, engine: E);
 }
